More tests and fixes. Test coverage now at 80%.

+++ lib/SdsDoctrineExtensions/AccessControl/Subscriber/AccessControl.php
@@ -6,9 +6,10 @@
  */
 namespace SdsDoctrineExtensions\AccessControl\Subscriber;
 
+use Doctrine\Common\Annotations\Reader;
 use Doctrine\Common\EventSubscriber;
-
 use Doctrine\ODM\MongoDB\Event\LifecycleEventArgs;
+use Doctrine\ODM\MongoDB\Event\LoadClassMetadataEventArgs;
 use Doctrine\ODM\MongoDB\Event\OnFlushEventArgs;
 use Doctrine\ODM\MongoDB\Events as ODMEvents;
 use SdsCommon\AccessControl\AccessControlledInterface;
@@ -19,15 +20,24 @@ use SdsCommon\User\RoleAwareUserInterface;
 use SdsDoctrineExtensions\AccessControl\AccessController;
 use SdsDoctrineExtensions\AccessControl\Constant\Action;
 use SdsDoctrineExtensions\AccessControl\Event\Events as AccessControlEvents;
+use SdsDoctrineExtensions\AccessControl\Mapping\MetadataInjector\AccessControl as MetadataInjector;
+use SdsDoctrineExtensions\AnnotationReaderAwareTrait;
+use SdsDoctrineExtensions\AnnotationReaderAwareInterface;
+use SdsDoctrineExtensions\State\Event\EventArgs as StateEventArgs;
+use SdsDoctrineExtensions\State\Event\Events as StateEvents;
 
 /**
  *
  * @since   1.0
  * @author  Tim Roediger <superdweebie@gmail.com>
  */
-class AccessControl implements EventSubscriber, ActiveUserAwareInterface
+class AccessControl implements
+    EventSubscriber,
+    AnnotationReaderAwareInterface,
+    ActiveUserAwareInterface
 {
     use ActiveUserAwareTrait;
+    use AnnotationReaderAwareTrait;
 
     /**
      *
@@ -53,24 +63,43 @@ class AccessControl implements EventSubscriber, ActiveUserAwareInterface
      */
     public function getSubscribedEvents(){
         return array(
-            ODMEvents::onFlush
+            ODMEvents::loadClassMetadata,
+            ODMEvents::onFlush,
+            StateEvents::onStateChange
         );
     }
 
     /**
      *
-     * @param \SdsCommon\AccessControl\RoleAwareUserInterface $activeUser
+     * @param \Doctrine\Common\Annotations\Reader $annotationReader
+     * @param \SdsCommon\User\RoleAwareUserInterface $activeUser
+     * @param boolean $controlCreate
+     * @param boolean $controlUpdate
+     * @param boolean $controlDelete
      */
     public function __construct(
+        Reader $annotationReader,
         RoleAwareUserInterface $activeUser,
         $controlCreate = true,
         $controlUpdate = true,
         $controlDelete = true
     ) {
+        $this->setAnnotationReader($annotationReader);
         $this->setActiveUser($activeUser);
         $this->controlCreate = $controlCreate;
         $this->controlUpdate = $controlUpdate;
         $this->controlDelete = $controlDelete;
+    }
+
+    /**
+     *
+     * @param \Doctrine\ODM\MongoDB\Event\LoadClassMetadataEventArgs $eventArgs
+     */
+    public function loadClassMetadata(LoadClassMetadataEventArgs $eventArgs)
+    {
+        $metadata = $eventArgs->getClassMetadata();
+        $metadataInjector = new MetadataInjector($this->annotationReader);
+        $metadataInjector->loadMetadataForClass($metadata);
     }
 
     /**
@@ -83,16 +112,24 @@ class AccessControl implements EventSubscriber, ActiveUserAwareInterface
         $unitOfWork = $documentManager->getUnitOfWork();
         $eventManager = $documentManager->getEventManager();
 
-        //Check create permissions
-        if ($this->controlCreate){
-            foreach ($unitOfWork->getScheduledDocumentInsertions() as $document) {
-                if($document instanceof AccessControlledInterface &&
-                    $document instanceof StateAwareInterface &&
+        foreach ($unitOfWork->getScheduledDocumentInsertions() as $document) {
+            if($document instanceof AccessControlledInterface) {
+
+                //Set stateEqualToParent on permissions
+                if ($document instanceof StateAwareInterface) {
+                    $documentState = $document->getState();
+                    foreach ($document->getPermissions() as $permission){
+                        $permission->setStateEqualToParent(($permission->getState() == $documentState));
+                    }
+                }
+
+                //Check create permissions
+                if ($this->controlCreate &&
                     !AccessController::isActionAllowed($document, Action::create, $this->activeUser)
                 ) {
                     //stop creation
                     $unitOfWork->detach($document);
-                    
+
                     if ($eventManager->hasListeners(AccessControlEvents::createDenied)) {
                         $eventManager->dispatchEvent(
                             AccessControlEvents::createDenied,
@@ -106,8 +143,38 @@ class AccessControl implements EventSubscriber, ActiveUserAwareInterface
         //Check update permissions
         if ($this->controlUpdate){
             foreach ($unitOfWork->getScheduledDocumentUpdates() as $document) {
-                if($document instanceof AccessControlledInterface &&
-                    $document instanceof StateAwareInterface &&
+
+                // Skip any updates on fields marked with @DoNotAccessControlUpdate
+                $changeSet = $unitOfWork->getDocumentChangeSet($document);
+                $metadata = $documentManager->getClassMetadata(get_class($document));
+                $checkPermission = false;
+                foreach ($changeSet as $field => $change) {
+                    if (!isset($metadata->fieldMappings[$field][MetadataInjector::doNotAccessControlUpdate])) {
+                        $checkPermission = true;
+                        break;
+                    } elseif (!$metadata->fieldMappings[$field][MetadataInjector::doNotAccessControlUpdate]) {
+                        $checkPermission = true;
+                        break;
+                    }
+                }
+                if (!$checkPermission){
+                    continue;
+                }
+
+                // allow updates to @stateField. If you need to control state updates, enable
+                // access control in the state extension
+                if ($document instanceof StateAwareInterface) {
+
+                    $changeSet = $unitOfWork->getDocumentChangeSet($document);
+                    $metadata = $documentManager->getClassMetadata(get_class($document));
+                    $field = $metadata->stateField;
+
+                    if (count($changeSet) == 1 && isset($changeSet[$field])) {
+                        continue;
+                    }
+                }
+
+                if ($document instanceof AccessControlledInterface &&
                     !AccessController::isActionAllowed($document, Action::update, $this->activeUser)
                 ) {
                     //stop updates
@@ -127,7 +194,6 @@ class AccessControl implements EventSubscriber, ActiveUserAwareInterface
         if ($this->controlDelete){
             foreach ($unitOfWork->getScheduledDocumentDeletions() as $document) {
                 if($document instanceof AccessControlledInterface &&
-                    $document instanceof StateAwareInterface &&
                     !AccessController::isActionAllowed($document, Action::delete, $this->activeUser)
                 ) {
                     //stop delete
@@ -140,6 +206,21 @@ class AccessControl implements EventSubscriber, ActiveUserAwareInterface
                         );
                     }
                 }
+            }
+        }
+    }
+
+    /**
+     *
+     * @param \SdsDoctrineExtensions\State\Event\EventArgs $eventArgs
+     */
+    public function onStateChange(StateEventArgs $eventArgs){
+        $document = $eventArgs->getDocument();
+
+        if($document instanceof AccessControlledInterface) {
+            $toState = $eventArgs->getToState();
+            foreach ($document->getPermissions() as $permission){
+                $permission->setStateEqualToParent(($permission->getState() == $toState));
             }
         }
     }
